refactor: modularize logging library structure
- Separate lib.rs into focused modules:
  - config.rs: LoggingConfig + ConfigPlugin + Environment
  - record.rs: LogRecord + LogLevel + helpers
  - loggers/: Logger trait + ConsoleLogger + JsonLogger
  - utils.rs: generate_request_id + formatting helpers
  - lib.rs: Re-exports + documentation + tests

- Update dependencies: add tyl-errors, tyl-config, serde_yaml
- Implement ConfigPlugin trait for tyl-config integration
- All 15 tests passing (9 unit + 4 integration + 2 doc)
- Maintain hexagonal architecture and full API compatibility
