+++ Cargo.toml
@@ -13,7 +13,12 @@ keywords = ["logging", "structured-logging", "hexagonal-architecture", "tyl", "f
 categories = ["development-tools::debugging", "rust-patterns"]
 
 [dependencies]
+# TYL Framework dependencies
+tyl-errors = { path = "../tyl-errors" }
+tyl-config = { path = "../tyl-config" }
+
+# Core dependencies  
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+serde_yaml = "0.9"
 uuid = { version = "1.0", features = ["v4"] }
-thiserror = "1.0"
